Removed models/equilibrium.py and moved fake_equilibrium methods to indica/equilibrium.py where they're better placed
